[Decomp][Player] Cohesion split: extract offline DB-lookup statics to PlayerDbLookup.cpp
