+++ src/game/Object/Player.cpp
@@ -4332,147 +4332,12 @@ uint32 Player::GetTalentResetCost()    const
     return resetTalentsCost(); // this function added in dev21 - remove this comment if this line works
 }
 
-/**
- * @brief Loads a player's guild identifier from the database.
- *
- * @param guid The player GUID to query.
- * @return The guild identifier, or zero if none is found.
- */
-uint32 Player::GetGuildIdFromDB(ObjectGuid guid)
-{
-    uint32 lowguid = guid.GetCounter();
-
-    QueryResult* result = CharacterDatabase.PQuery("SELECT `guildid` FROM `guild_member` WHERE `guid`='%u'", lowguid);
-    if (!result)
-    {
-        return 0;
-    }
-
-    uint32 id = result->Fetch()[0].GetUInt32();
-    delete result;
-    return id;
-}
-
-ObjectGuid Player::GetGuildGuidFromDB(ObjectGuid guid)
-{
-    if (uint32 guildId = GetGuildIdFromDB(guid))
-    {
-        return ObjectGuid(HIGHGUID_GUILD, GetGuildIdFromDB(guid));
-    }
-    else
-    {
-        return ObjectGuid();
-    }
-}
-
-/**
- * @brief Loads a player's guild rank from the database.
- *
- * @param guid The player GUID to query.
- * @return The guild rank, or zero if none is found.
- */
-uint32 Player::GetRankFromDB(ObjectGuid guid)
-{
-    QueryResult* result = CharacterDatabase.PQuery("SELECT `rank` FROM `guild_member` WHERE `guid`='%u'", guid.GetCounter());
-    if (result)
-    {
-        uint32 v = result->Fetch()[0].GetUInt32();
-        delete result;
-        return v;
-    }
-    else
-    {
-        return 0;
-    }
-}
-
 void Player::SendGuildDeclined(std::string name, bool autodecline)
 {
     WorldPacket data(SMSG_GUILD_DECLINE, 10);
     data << name;
     data << uint8(autodecline);
     GetSession()->SendPacket(&data);
-}
-
-
-uint32 Player::GetArenaTeamIdFromDB(ObjectGuid guid, ArenaType type)
-{
-    QueryResult* result = CharacterDatabase.PQuery("SELECT `arena_team_member`.`arenateamid` FROM `arena_team_member` JOIN `arena_team` ON `arena_team_member`.`arenateamid` = `arena_team`.`arenateamid` WHERE `guid`='%u' AND `type`='%u' LIMIT 1", guid.GetCounter(), type);
-    if (!result)
-    {
-        return 0;
-    }
-
-    uint32 id = (*result)[0].GetUInt32();
-    delete result;
-    return id;
-}
-
-/**
- * @brief Loads or derives a player's saved zone identifier from the database.
- *
- * @param guid The player GUID to query.
- * @return The resolved zone identifier, or zero on failure.
- */
-uint32 Player::GetZoneIdFromDB(ObjectGuid guid)
-{
-    uint32 lowguid = guid.GetCounter();
-    QueryResult* result = CharacterDatabase.PQuery("SELECT `zone` FROM `characters` WHERE `guid`='%u'", lowguid);
-    if (!result)
-    {
-        return 0;
-    }
-    Field* fields = result->Fetch();
-    uint32 zone = fields[0].GetUInt32();
-    delete result;
-
-    if (!zone)
-    {
-        // stored zone is zero, use generic and slow zone detection
-        result = CharacterDatabase.PQuery("SELECT `map`,`position_x`,`position_y`,`position_z` FROM `characters` WHERE `guid`='%u'", lowguid);
-        if (!result)
-        {
-            return 0;
-        }
-        fields = result->Fetch();
-        uint32 map = fields[0].GetUInt32();
-        float posx = fields[1].GetFloat();
-        float posy = fields[2].GetFloat();
-        float posz = fields[3].GetFloat();
-        delete result;
-
-        zone = sTerrainMgr.GetZoneId(map, posx, posy, posz);
-
-        if (zone > 0)
-        {
-            CharacterDatabase.PExecute("UPDATE `characters` SET `zone`='%u' WHERE `guid`='%u'", zone, lowguid);
-        }
-    }
-
-    return zone;
-}
-
-/**
- * @brief Loads a player's level from the database.
- *
- * @param guid The player GUID to query.
- * @return The stored level, or zero on failure.
- */
-uint32 Player::GetLevelFromDB(ObjectGuid guid)
-{
-    uint32 lowguid = guid.GetCounter();
-
-    QueryResult* result = CharacterDatabase.PQuery("SELECT `level` FROM `characters` WHERE `guid`='%u'", lowguid);
-    if (!result)
-    {
-        return 0;
-    }
-
-    Field* fields = result->Fetch();
-    uint32 level = fields[0].GetUInt32();
-    delete result;
-
-    return level;
 }
 
 /**
